refactor: ♻️Modify folder structure
